pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = ""
authors = ["Benoit Baqué de Sariac <benoit.bds@gmail.com>"]
readme = "README.md"
packages = [{ include = "orchestrator" }]

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.5.3"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
langchain-openai = "^0.3.28"
python-dotenv = "^1.1.1"
typer = "^0.16.0"
pydantic = "^2.11.7"
langchain = ">=0.2.0"
httpx = "^0.28.1"
python-multipart = "^0.0.9"
pymupdf = "^1.25.1"
python-docx = "^1.1.2"
pytesseract = "^0.3.13"
pillow = "^10.4.0"
tiktoken = "^0.7.0"
numpy = "^2.1.3"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
ruff = "^0.12.3"
httpx-ws = "^0.7.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
